docs: Swift-doc 주석 작성
