+++ AppProduct/AppProduct/Core/Common/Error/Types/AuthError.swift
