+++ AppProduct/AppProduct/Core/Common/Error/Types/ErrorSeverity.swift
